Refactor DateWrapper to use Date property instead of _date
Renamed the internal field `_date` to `Date` in the `DateWrapper` class and updated all references accordingly. Updated the `DateWrapper` class's constructor, methods, and the `IsNullOrEmpty`, `Equals`, and `GetHashCode` methods to use the new `Date` property. Modified the XML serialization and deserialization methods to use the `Date` property. Updated `DateWrapperSerializableTest` and `DateWrapperTest` classes to reflect these changes. Ensured consistent use of the `Date` property throughout the codebase.